Add the react-router-dom + one route that will be used to share the link with other people in read-only mode.
