+++ package.json
@@ -8,6 +8,7 @@
     "@fortawesome/free-solid-svg-icons": "^5.12.0",
     "@fortawesome/react-fontawesome": "^0.1.8",
     "@types/localforage": "0.0.34",
+    "@types/react-router-dom": "^5.1.3",
     "@types/uuid": "^3.4.6",
     "axios": "^0.19.0",
     "axios-retry": "^3.1.2",
@@ -26,6 +27,7 @@
     "react-image-file-resizer": "^0.2.1",
     "react-intl": "^3.9.2",
     "react-modal": "^3.11.1",
+    "react-router-dom": "^5.1.2",
     "react-scripts": "3.3.0",
     "react-switch": "^5.0.1",
     "react-transition-group": "^4.3.0",
