+++ src/components/MainPanel/MainPanel.css
